requirements.txt
numpy
pillow
matplotlib
tqdm
pandas
einops
hydra-core
lightning
pytorch_lightning
websockets
opencv-python
wandb
cvxpy
transformers
pycocoevalcap
torchmetrics
openai
salesforce-lavis
timm==0.9
git+https://github.com/locuslab/qpth.git
git+https://github.com/facebookresearch/detectron2.git